refactor: moved Skeleton
