+++ apps/rtk-query/src/shared/components/AudioPlayer/AudioPlayerSceleton/Skeleton.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
-import { Card } from '../Card'
-import { Typography } from '../Typography'
-import { Skeleton } from './Skeleton'
+import { Card } from '../../Card'
+import { Typography } from '../../Typography'
+import { Skeleton } from './Skeleton.tsx'
 
 const meta = {
   title: 'Components/Skeleton',
