+++ apps/rtk-query/src/shared/components/AudioPlayer/AudioPlayerSceleton/Skeleton.tsx
